refactor: make WindowId an isize type
To reduce the number of casts

(cherry picked from commit 555c3a412195d25ad95447ea21ab3807d076ea48)

+++ native/desktop-macos/headers/desktop_macos.h
@@ -76,7 +76,7 @@ typedef struct NativeApplicationConfig {
   bool disable_character_palette_menu_item;
 } NativeApplicationConfig;
 
-typedef int64_t NativeWindowId;
+typedef intptr_t NativeWindowId;
 
 typedef uintptr_t NativeKeyModifiersSet;
